requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
SQLAlchemy==2.0.23
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pandas==2.2.3